Move levelling system commands from main bot script to cog
